requirements.txt
yt-dlp
google-auth==2.29.0
google-api-python-client==2.118.0
anthropic==0.40.0
python-dotenv==1.0.1
flask==3.1.0
flask-cors==4.0.1
